Rename `Algo` folder to `Algorithm`
