+++ include/Algorithm/AlgorithmBase.h
