+++ include/Algorithm/Delay/Fractional.h
@@ -24,7 +24,7 @@
 #include <cmath>
 #include <vector>
 
-#include "Algo/AlgorithmBase.h"
+#include "Algorithm/AlgorithmBase.h"
 #include "Util/FloatingPointConcept.h"
 
 namespace Clover::Delay {
