+++ include/Algorithm/DynamicRange/UpwardDCCompression.h
@@ -23,7 +23,7 @@
 #include <array>
 #include <cmath>
 
-#include "Algo/AlgorithmBase.h"
+#include "Algorithm/AlgorithmBase.h"
 #include "Util/FloatingPointConcept.h"
 
 namespace Clover::DynamicRange {
