+++ include/Algorithm/Envelope/Basic.h
@@ -23,7 +23,7 @@
 #include <assert.h>
 #include <cmath>
 
-#include "Algo/AlgorithmBase.h"
+#include "Algorithm/AlgorithmBase.h"
 #include "Util/FloatingPointConcept.h"
 
 namespace Clover::Envelope {
