+++ include/Algorithm/Envelope/DC.h
@@ -20,7 +20,7 @@
  *
  */
 
-#include "Algo/AlgorithmBase.h"
+#include "Algorithm/AlgorithmBase.h"
 #include "Util/FloatingPointConcept.h"
 
 namespace Clover::Envelope {
