+++ include/Algorithm/Stereo/Pan.h
@@ -23,7 +23,7 @@
 #include <algorithm>
 #include <array>
 
-#include "Algo/AlgorithmBase.h"
+#include "Algorithm/AlgorithmBase.h"
 #include "Util/Calc.h"
 #include "Util/FloatingPointConcept.h"
