+++ include/Algorithm/Wavetable/OscillatorInterface.h
