+++ include/Nodes/Envelope/ADSR.h
@@ -20,7 +20,7 @@
  *
  */
 
-#include "Algo/Envelope/ADSR.h"
+#include "Algorithm/Envelope/ADSR.h"
 #include "Graph/AudioNode.h"
 #include "Nodes/StepSequencer/lib.h"
