+++ include/Nodes/Stereo/Pan.h
@@ -20,7 +20,7 @@
  *
  */
 
-#include "Algo/Stereo/Pan.h"
+#include "Algorithm/Stereo/Pan.h"
 #include "Graph/AudioFrame.h"
 #include "Graph/AudioNode.h"
 #include "Util/Calc.h"
